+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- limited support for whole pattern recursion (`\g<0>`)
+
+### Fixed
+
+- SystemStackError with recursive group calls (`(a|b\g<1>)c`)
+
 ## [3.8.0] - 2022-09-25
 
 ### Added